requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
itsdangerous==2.1.2
werkzeug==3.0.1
